requirements.txt
pyyaml
sentry-sdk